Move component resolvers
