+++ pkg/components/cnudie/factory.go
@@ -5,6 +5,8 @@ import (
 	"fmt"
 	"os"
 
+	"github.com/gardener/landscaper/pkg/components/cnudie/oci"
+
 	"github.com/gardener/component-cli/ociclient"
 	"github.com/gardener/component-cli/ociclient/cache"
 	"github.com/gardener/component-cli/ociclient/credentials"
@@ -25,7 +27,6 @@ import (
 	"github.com/gardener/landscaper/pkg/components/cnudie/helmrepo"
 	"github.com/gardener/landscaper/pkg/components/model"
 	"github.com/gardener/landscaper/pkg/components/model/types"
-	componentsregistry "github.com/gardener/landscaper/pkg/landscaper/registry/components"
 	"github.com/gardener/landscaper/pkg/utils"
 )
 
@@ -43,13 +44,13 @@ func (*Factory) NewRegistryAccess(ctx context.Context,
 
 	logger, _ := logging.FromContextOrNew(ctx, nil)
 
-	compResolver, err := componentsregistry.New(sharedCache)
+	compResolver, err := oci.New(sharedCache)
 	if err != nil {
 		return nil, fmt.Errorf("unable to create component registry manager: %w", err)
 	}
 
 	if localRegistryConfig != nil {
-		localRegistry, err := componentsregistry.NewLocalClient(localRegistryConfig.RootPath)
+		localRegistry, err := oci.NewLocalClient(localRegistryConfig.RootPath)
 		if err != nil {
 			return nil, err
 		}
@@ -81,7 +82,7 @@ func (*Factory) NewRegistryAccess(ctx context.Context,
 		return nil, err
 	}
 
-	componentsOCIRegistry, err := componentsregistry.NewOCIRegistryWithOCIClient(logger, ociClient, inlineCd)
+	componentsOCIRegistry, err := oci.NewOCIRegistryWithOCIClient(logger, ociClient, inlineCd)
 	if err != nil {
 		return nil, err
 	}
@@ -134,7 +135,7 @@ func (*Factory) NewOCIRegistryAccess(ctx context.Context,
 		return nil, err
 	}
 
-	componentResolver, err := componentsregistry.NewOCIRegistryWithOCIClient(log, ociClient, predefinedComponentDescriptors...)
+	componentResolver, err := oci.NewOCIRegistryWithOCIClient(log, ociClient, predefinedComponentDescriptors...)
 	if err != nil {
 		return nil, err
 	}
@@ -157,7 +158,7 @@ func (*Factory) NewOCIRegistryAccessFromDockerAuthConfig(ctx context.Context,
 		return nil, err
 	}
 
-	componentResolver, err := componentsregistry.NewOCIRegistryWithOCIClient(log, ociClient, predefinedComponentDescriptors...)
+	componentResolver, err := oci.NewOCIRegistryWithOCIClient(log, ociClient, predefinedComponentDescriptors...)
 	if err != nil {
 		return nil, errors.Wrap(err, "unable to setup components registry")
 	}
@@ -223,7 +224,7 @@ func (*Factory) NewOCITestRegistryAccess(address, username, password string) (mo
 }
 
 func (*Factory) NewLocalRegistryAccess(rootPath string) (model.RegistryAccess, error) {
-	localComponentResolver, err := componentsregistry.NewLocalClient(rootPath)
+	localComponentResolver, err := oci.NewLocalClient(rootPath)
 	if err != nil {
 		return nil, err
 	}
