+++ pkg/components/cnudie/oci/oci.go
@@ -1,8 +1,4 @@
-// SPDX-FileCopyrightText: 2020 SAP SE or an SAP affiliate company and Gardener contributors.
-//
-// SPDX-License-Identifier: Apache-2.0
-
-package componentsregistry
+package oci
 
 import (
 	"context"
