+++ pkg/components/cnudie/oci/testdata/comp1/component-descriptor.yaml
