+++ pkg/components/cnudie/oci/testdata/comp1/config.json
