+++ pkg/components/registries/registries_test.go
@@ -7,6 +7,8 @@ package registries
 import (
 	"context"
 
+	"github.com/gardener/landscaper/pkg/components/cnudie/oci"
+
 	cdv2 "github.com/gardener/component-spec/bindings-go/apis/v2"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
@@ -14,7 +16,6 @@ import (
 	lsv1alpha1 "github.com/gardener/landscaper/apis/core/v1alpha1"
 	"github.com/gardener/landscaper/pkg/components/model"
 	"github.com/gardener/landscaper/pkg/components/model/types"
-	componentsregistry "github.com/gardener/landscaper/pkg/landscaper/registry/components"
 )
 
 var (
@@ -28,7 +29,7 @@ var _ = BeforeSuite(func() {
 	registryAccess, err = NewFactory().NewLocalRegistryAccess("./testdata/registry")
 	Expect(err).ToNot(HaveOccurred())
 
-	repository := componentsregistry.NewLocalRepository("./testdata/registry")
+	repository := oci.NewLocalRepository("./testdata/registry")
 	repositoryContext, err = cdv2.NewUnstructured(repository)
 	Expect(err).ToNot(HaveOccurred())
 })
