+++ pkg/landscaper/installations/executions/executions_suite_test.go
@@ -10,14 +10,14 @@ import (
 	"path/filepath"
 	"testing"
 
+	"github.com/gardener/landscaper/pkg/components/cnudie/oci"
+
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/gardener/landscaper/test/utils/envtest"
-
-	componentsregistry "github.com/gardener/landscaper/pkg/landscaper/registry/components"
 
 	testutils "github.com/gardener/landscaper/test/utils"
 )
@@ -47,7 +47,7 @@ var _ = AfterSuite(func() {
 
 func createDefaultContextsForNamespace(kubeClient client.Client) {
 	// create default repo for all namespaces
-	repoCtx := componentsregistry.NewLocalRepository("../testdata/registry")
+	repoCtx := oci.NewLocalRepository("../testdata/registry")
 	list, err := os.ReadDir("./testdata/state")
 	Expect(err).To(Succeed())
 	for _, d := range list {
